moved classes related to InfoPanel into their own package
